requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coverage==5.3.1
cryptography==3.3.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.4
django-allauth==0.44.0
django-debug-toolbar==3.2
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
